Add chart

+++ tickers/src/main/resources/db/migration/V1_2__buyticker.sql
@@ -1,6 +1,6 @@
 CREATE TABLE BUY_TICKER (
                         ID serial NOT NULL,
-                        VALOR BIGINT NOT NULL,
+                        VALOR double precision NOT NULL,
                         DATA TIMESTAMP NOT NULL,
                         QUANTIDADE double precision NOT NULL,
                         TICKER TEXT NOT NULL);
